Make images scale properly within masonry layout
I do not think that the images need to be wrapped in a Card,
but it is convenient to have them so wrapped, so that I can
add drag handles, checkboxes, titles, logos, etc etc.

The slider's variable is plumbed through to the image drawing
component to force image size updates when the surrounding
Card is resized by the number of masonry columns getting
updated.

This all runs a bit slow, and sometimes we can see images
resizing, so it is clunky on both counts.  I am sure that there
must be a better way.

Uses code derived from Kirill Novikov's avatar component, see
comment in source.