Allow removal of images from the search
Includes renaming 'Detail search' button to 'Edit search' --
which hints at the new aspect of the functionality, but
unfortately loses the hint of the most interesting aspect
of the functionality.

At this point, treating the primary input as special seems to cease
to make sense -- surely it should just be a member of the same
array as everything else. However, there is not time to try to
engineer this properly.