Rework input editing: step 2
Make the selections appear to the right.

The layout on the RHS could be a lot better.

Would like to have the current image remain selected when
we press 'Update selections' but this causes the overlay
on the LHS to shift to the background image of the current
selection -- but only for new bits of the image. Can work
around that by continuing to set the detailImage to undefined,
so just do that.